+++ src/modules/vtex/install/install.ts
@@ -7,9 +7,13 @@ import { ContentManifest } from '../../../shared/models/global';
 import fs from 'fs';
 import { runMultipleCommand } from '../../../shared/util/run-multiple-command';
 const directory = process.cwd() + '/';
+let commandAdminPackage: string = '';
 
-export default async function() {
+
+export default  async (adminPackage: string) => {
   log.info('Install dependencies');
+
+  commandAdminPackage = adminPackage;
   //1. Busco en el directorio actual si existen proyectos con el archivo manifest.json
   let response = await findProject([directory]);
   // si no se encontraron projectos en el directorio actual, paso a buscar en los sub directorios.
@@ -50,11 +54,11 @@ const findProject = async (files: Array<string>) => {
 const findNodeOrReactFolder = async (file: string) => {
   if (fs.existsSync(file + '/node') == true) {
     log.info('Proyect contain the folder node');
-    const command = `cd ${file}/node && npm install && cd ../..`;
+    const command = `cd ${file}/node && ${commandAdminPackage} install && cd ../..`;
     runMultipleCommand(command);
   } else if (fs.existsSync(file + '/react') == true) {
     log.info('Proyect contain the folder react');
-    const command = `cd ${file}/react && npm install && cd ../..`;
+    const command = `cd ${file}/react && ${commandAdminPackage} install && cd ../..`;
     runMultipleCommand(command);
   }
 };